Fix for mixed precision support

+++ source/slideflow/gan/semantic/discriminator.py
@@ -48,5 +48,5 @@ def create_discriminator(image_size, df_dim=64):
 	h5_act = tf.keras.layers.LeakyReLU()(h5)
 	#h6 = tf.keras.layers.Lambda(lambda x: tf.reduce_sum(x, axis=[1,2]))(h5_act)
 	h6 = tf.keras.layers.Flatten()(h5_act)
-	output = DenseSN(1)(h6)
+	output = DenseSN(1, dtype=tf.float32)(h6)
 	return tf.keras.models.Model(image, output)
